+++ gradabeam/gradabeam_reference.py
@@ -107,7 +107,7 @@ class GradaBeamReference:
         assert isinstance(start_sequence, str)
         seed_node = RolloutNode(
             seq=start_sequence,
-            fitness=np.float32(0.0),
+            fitness=0.0,
             edits_since_root=0,
             probs=None,
             pos_and_chars=None,
@@ -374,7 +374,7 @@ class GradaBeamReference:
         return [
             RolloutNode(
                 seq=seq,
-                fitness=float(f),  # type: ignore[arg-type]
+                fitness=float(f),
                 probs=probs,
                 edits_since_root=n.edits_since_root + int(num_edits),
                 pos_and_chars=n.pos_and_chars,
